requirements.txt
flake8
psycopg2-binary
pytest
pytest-mock
